requirements.txt
attr==0.3.2
colorama==0.4.6
ConfigParser==7.2.0
cryptography==46.0.3
discord.py==2.6.4
discord-ext-voice-recv>=0.4.0
docutils==0.22.4
faster_whisper==1.2.1
filelock==3.20.1
HTMLParser==0.0.2
ipython==9.8.0
ipywidgets==8.1.8
keyring==25.7.0
numpy==2.4.0
Pillow==12.0.0
protobuf==6.33.2
pyOpenSSL==25.3.0
python-dotenv==1.2.1
PyYAML==6.0.1
pyttsx3==2.99
redis==7.1.0
scipy==1.16.3
xai-sdk>=0.3.0
Sphinx==9.0.4
thread==2.0.6
urllib3_secure_extra==0.1.0